Add retry utility and use in connectors
