+++ connectors/sx.py
@@ -5,30 +5,13 @@ from aiohttp import ClientSession
 from aiohttp.client_exceptions import ClientError
 from typing import Any
 
+from utils.retry import retry
+
 API_REST = "https://api.sx.bet"
 
 
 class SxError(Exception):
     """Raised when SX API responses are invalid."""
-
-
-def retry(attempts: int = 3, delay: float = 1.0):
-    def decorator(func):
-        async def wrapper(*args, **kwargs):
-            for i in range(attempts):
-                try:
-                    return await func(*args, **kwargs)
-                except Exception as exc:
-                    logging.debug(
-                        "Attempt %s/%s failed: %s", i + 1, attempts, exc, exc_info=True
-                    )
-                    if i == attempts - 1:
-                        raise
-                    await asyncio.sleep(delay)
-
-        return wrapper
-
-    return decorator
 
 
 @retry()
